# openAPI/robot_api.yaml
openapi: 3.0.2
info:
  title: RoboCupMS - Robot API
  description: API umožnuje vytváření, modifikaci, registraci robotů. Robot je vytvářen na registraci tymu v soutěži.
  version: 1.0.0
  contact:
    email: m1_krcma@utb.cz
servers:
  - url: https://utb-robogames.com:8080/api/teamRegistration
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-TOKEN
paths:
  /get:
    get:
      summary: Navrátí robota s konkretním ID
      parameters:
        - in: query
          name: id
          required: true
          description: ID robota
          schema:
            type: string
      responses:
        "200":
          description: Robot
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      number:
                        type: integer
                        example: 34
                      name:
                        type: string
                        example: "Robot"
                      confirmed:
                        type: boolean
                        example: false
                      category:
                        type: string
                        example: "OPEN"
                      teamRegistrationID:
                        type: integer
                        example: 10
                      teamName:
                        type: string
                        example: "Myteam"
                      disciplineID:
                        type: integer
                        example: 4
                      diciplineName:
                        type: string
                        example: "Mini Sumo"

  /all:
    get:
      summary: Navrátí všechny roboty vytvořene pro registraci tymu v určitem ročniku soutěže
      description: Navrátí všechny vytvořene roboty pro danou registracu týmu, ve kterém se přihlášený uživatel nachází.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: year
          required: true
          description: Ročnik soutěže, do které je tým registrovaný
          schema:
            type: integer
      responses:
        "200":
          description: Robot
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        number:
                          type: integer
                          example: 34
                        name:
                          type: string
                          example: "Robot"
                        confirmed:
                          type: boolean
                          example: false
                        category:
                          type: string
                          example: "OPEN"
                        teamRegistrationID:
                          type: integer
                          example: 10
                        teamName:
                          type: string
                          example: "Myteam"
                        disciplineID:
                          type: integer
                          example: 4
                        diciplineName:
                          type: string
                          example: "Mini Sumo"

  /allConfirmed:
    get:
      summary: Navrátí všechny roboty s potvrzenou registrací pro určity ročník soutěže
      parameters:
        - in: query
          name: year
          required: true
          description: Ročnik soutěže
          schema:
            type: integer
      responses:
        "200":
          description: Robot
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        number:
                          type: integer
                          example: 34
                        name:
                          type: string
                          example: "Robot"
                        confirmed:
                          type: boolean
                          example: false
                        category:
                          type: string
                          example: "OPEN"
                        teamRegistrationID:
                          type: integer
                          example: 10
                        teamName:
                          type: string
                          example: "Myteam"
                        disciplineID:
                          type: integer
                          example: 4
                        diciplineName:
                          type: string
                          example: "Mini Sumo"

  /create:
    post:
      summary: Vytvoří nového robata
      description: Vytvořit robota může jakýkoli člen týmu. Robot je výtvářen na registraci týmu v určitém ročníku soutěže.
      parameters:
        - in: query
          name: year
          required: true
          description: Ročník soutěže
          schema:
            type: integer
        - in: query
          name: name
          required: true
          description: Jméno robota (uníkátní v rámci ročníku soutěže)
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Odpověd o stavu provedeného požadavku
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: string
                    example: "success"

  /rename:
    put:
      summary: Změní jméno robota
      description: Změní jméno robota může jakýkoli člen týmu. Jméno musí být v rámci ročníku unikátní. Pokud již
      parameters:
        - in: query
          name: year
          required: true
          description: Ročník soutěže
          schema:
            type: integer
        - in: query
          name: id
          required: true
          description: ID robota, kterému chceme změnit jméno
          schema:
            type: integer
        - in: query
          name: name
          required: true
          description: Nové jméno robota (uníkátní v rámci ročníku soutěže)
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Odpověd o stavu provedeného požadavku
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: string
                    example: "success"

  /register:
    put:
      summary: Registruje robota do určité disciplíny
      description: Registrovat robota může jakýkoli člen týmu. Robota je možné přihlásit dokud soutěž nezačala.
      parameters:
        - in: query
          name: robotID
          required: true
          description: ID robota
          schema:
            type: integer
        - in: query
          name: disciplineID
          required: true
          description: ID disciplíny, do které robota registrujeme
          schema:
            type: integer
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Odpověd o stavu provedeného požadavku
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: string
                    example: "success"

  /confirmRegistration:
    put:
      summary: Povrdí/nepovrdí registraci robota v disciplíně
      description: Mají přístup jen uživatelé s touto rolí [ADMIN, LEADER, ASSISTANT].
      parameters:
        - in: query
          name: id
          required: true
          description: ID robota
          schema:
            type: integer
        - in: query
          name: confirmed
          required: true
          description: Registrace potvrzena/nepotvrzena
          schema:
            type: boolean
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Odpověd o stavu provedeného požadavku
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: string
                    example: "success"


  /unregister:
    delete:
      summary: Zruší registraci existujícího robota
      description: Zrušit registraci robota může jakýkoli člen týmu. Zrušení registrace nebude možne pokud již byla povrzena.
      parameters:
        - in: query
          name: robotID
          required: true
          description: ID robota
          schema:
            type: integer
        - in: query
          name: disciplineID
          required: true
          description: ID disciplíny, do které robota registrujeme
          schema:
            type: integer
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Odpověd o stavu provedeného požadavku
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: string
                    example: "success"

  /remove:
    delete:
      summary: Odstraní robota
      description: Odstranit robota může jakýkoli člen týmu. Odstranění robota nebude možné ve chvíli, kdy již jeho registrace do discipliny byla povtrzena.
      parameters:
        - in: query
          name: year
          required: true
          description: Ročník soutěže
          schema:
            type: integer
        - in: query
          name: id
          required: true
          description: ID robota, který bude odstraněn
          schema:
            type: integer
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Odpověd o stavu provedeného požadavku
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: "RESPONSE"
                  data:
                    type: string
                    example: "success"